Switched to "Member-SUBM" spec status and added right boilerplate to SoTD
ReSpec has limited support for Member Submissions. The "Member-SUBM"
specification status it defines does not really do the job properly. It's
still better than sticking to the "unofficial" status.

I added the right boilerplate to the "Status of This Document" section and
the right document license for the submission. I also removed myself from
the list of editors (mostly to streamline the submission process).

There remain a couple of things to change manually in the snapshot of the
specification that will need to be submitted:

1. Drop the useless empty "Previous Version" item before the list of editors
2. Remove all the paragraphs after "Please consult the complete list of
acknowledged W3C Member Submissions" in the Status of This Document section.
These paragraphs are incorrectly added by ReSpec.