Added an explicit note that semantic enrichment needs further work
Turning an Extended Web Thing into a Semantic Web Thing is alluded to in the
spec but not really explained (on top of "you may use JSON-LD). The note
clarifies that this would require further work which could potentially affect
the design of the data model.

The note links to the QUTD set of ontologies as example.